# app/docs/swagger.yaml
basePath: /v1
definitions:
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.OTPInput:
    properties:
      token:
        type: string
      user_id:
        type: string
    type: object
  model.OkGenerateOTP:
    properties:
      base32: {}
      otp-auth_url: {}
    type: object
  model.OkLoginResponce:
    properties:
      email:
        type: string
      otp_enabled:
        type: boolean
      refresh_token:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  model.OkResponse:
    properties:
      message:
        type: string
    type: object
...
  model.Track:
    properties:
      album:
        example: Album name
        type: string
      album_artist:
        example: Album artist name
        type: string
      artist:
        example: Artist name
        type: string
      bitrate:
        example: 320
        type: integer
      comment:
        example: Additional comments
        type: string
      composer:
        example: Composer name
        type: string
      disc:
        example: 1
        type: integer
      disc_total:
        example: 2
        type: integer
      duration:
        $ref: '#/definitions/time.Duration'
      genre:
        example: Genre name
        type: string
      lyrics:
        example: Lyrics of the track
        type: string
      sample_rate:
        example: 44100
        type: integer
      title:
        example: Title name
        type: string
      track:
        example: 3
        type: integer
      track_total:
        example: 10
        type: integer
      year:
        example: 2022
        type: integer
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - 1
    - 1000
    - 1000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - Nanosecond
    - Microsecond
    - Millisecond
...
host: s3streammedia.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a s3 media streamer server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: S3 Media Streamer Application API
  version: 0.0.1
paths:
  /audio/{playlist_id}:
    get:
      consumes:
      - '*/*'
      description: Streams audio_handler files in the specified directory as MP3 or
        FLAC.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        type: string
      - description: Control operation playlist_handler play
        in: path
        name: control
        type: string
      produces:
      - application/x-mpegURL
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Track'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Stream audio_handler files.
      tags:
      - track_handler-controller
  /audio/stream/{segment}:
    get:
      consumes:
      - '*/*'
      description: Streams audio_handler files in the specified directory as MP3 or
        FLAC.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        type: string
      - description: Control operation playlist_handler play
        in: path
        name: control
        type: string
      produces:
      - audio_handler/mpeg
      - audio_handler/flac
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Track'
            type: array
        "404":
          description: Segment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "406":
          description: Segment not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Stream audio_handler files.
      tags:
      - track_handler-controller
...
  /job/status:
    get:
      consumes:
      - application/json
      description: Check if the application server is running jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: All Job status
      tags:
      - health-controller
  /otp/disable:
    post:
      consumes:
      - application/json
      description: Disable OTP for a user_handler by setting 'otp_enabled' to 'false'
        in the database.
      parameters:
      - description: OTP input data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OTP disabled successfully
          schema:
            $ref: '#/definitions/model.OkLoginResponce'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to update OTP status
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disable OTP for a user_handler.
      tags:
      - OTP
  /otp/generate:
    post:
      consumes:
      - application/json
      description: Generate an OTP token for a user_handler and store it in the database.
      parameters:
      - description: OTP input data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OTP generated successfully
          schema:
            $ref: '#/definitions/model.OkGenerateOTP'
        "400":
          description: Invalid refresh payload
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Failed to find user_handler or invalid email/password
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to update OTP secret or URL
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate OTP for a user_handler.
      tags:
      - OTP
  /otp/validate:
    post:
      consumes:
      - application/json
      description: Validates a One-Time Password (OTP) for a user_handler.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: OTP Input
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OTP Valid
          schema:
            $ref: '#/definitions/model.OkResponse'
        "400":
          description: Bad Request - Invalid OTP
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Validates a One-Time Password (OTP).
      tags:
      - OTP
  /otp/verify:
    post:
      consumes:
      - application/json
      description: Verify the OTP token for a user_handler and update 'otp_enabled'
        and 'otp_verified' fields in the database.
      parameters:
      - description: OTP input data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.OTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified successfully
          schema:
            $ref: '#/definitions/model.OkLoginResponce'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Failed to update OTP status
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Verify OTP for a user_handler.
      tags:
      - OTP
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"