# app/handlers/REST/jobs_handler/main.go
package jobs_handler

import (
	"net/http"

	"github.com/bamzi/jobrunner"
	"github.com/gin-gonic/gin"
	"go.opentelemetry.io/otel"
)

type JobServiceInterface interface {
}

type JobHandler struct {
}

func NewJobHandler() *JobHandler {
	return &JobHandler{}
}

// JobStatus godoc
// @Summary All Job status
// @Description Check if the application server is running jobs
// @Tags health-controller
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{} "OK"
// @Failure 404 {object} map[string]string "Not Found"
// @Router /job/status [get]
func (h *JobHandler) JobStatus(c *gin.Context) {
	_, span := otel.Tracer("").Start(c.Request.Context(), "JobStatus")
	defer span.End()
	c.JSON(http.StatusOK, jobrunner.StatusJson())
}